% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compile_ggconf}
\alias{compile_ggconf}
\title{the core function of ggconf}
\usage{
compile_ggconf(cmd = "gg mtcars + p wt wt")
}
\arguments{
\item{cmd}{A character

compile_ggconf returns a built ggplot object.}
}
\description{
the core function of ggconf
}

