% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{find_first_index}
\alias{find_first_index}
\title{define constant values used in ggconf}
\usage{
find_first_index(pattern = "sz", table = c("x", "y", "size", "shape",
  "colour", "fill", "alpha", "stroke"), show_warn = TRUE, debug = FALSE)
}
\arguments{
\item{pattern}{the pattern string}

\item{table}{a table of string to be matched}

\item{show_warn}{boolean for showing ambiguous match warning}

\item{debug}{show debugging message if true}
}
\description{
define constant values used in ggconf
}

