% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{get_element_tree_clone}
\alias{get_element_tree_clone}
\title{get element tree clone}
\usage{
get_element_tree_clone()
}
\description{
{
   devtools::check() add a note about using ggplot2:::.element_tree
   because it is an internal object of other packages.
   Thus, a quick-and-dirty solution,
   I just copied the resulted data frame here.
}
}

