% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{get_theme_elem_name_conf}
\alias{get_theme_elem_name_conf}
\title{get all theme element configurations}
\usage{
get_theme_elem_name_conf(class = "element_text")
}
\arguments{
\item{class}{one of "element_text", "element_blank", "element_line", or
"element_rect"}
}
\description{
get all theme element configurations
}
