% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{ggconf_dbgmsg}
\alias{ggconf_dbgmsg}
\title{display ggconf debug message}
\usage{
ggconf_dbgmsg(...)
}
\arguments{
\item{...}{a sequence of objects passed to message()}
}
\description{
display ggconf debug message
}

