% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{remove_element_whatever}
\alias{remove_element_whatever}
\title{remove element_* function calls}
\usage{
remove_element_whatever(input = "theme(l=element_text(sz=20))")
}
\arguments{
\item{input}{An input string passed from users}
}
\description{
remove element_* function calls
}

