% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{replace_marks}
\alias{replace_marks}
\title{replace some marks for later ggconf parsing}
\usage{
replace_marks(input = "theme2(l.txt(size=12))")
}
\arguments{
\item{input}{An input string passed from users}
}
\description{
replace some marks for later ggconf parsing
}
