% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compile_ggconf}
\alias{compile_ggconf}
\title{the core function of ggconf}
\usage{
compile_ggconf(cmd = "")
}
\arguments{
\item{cmd}{preprocessed characters

compile_ggconf returns a built theme() object as string.}
}
\description{
the core function of ggconf
}
